+++ front-end/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {BrowserRouter as Router, Route, Switch} from "react-router-dom";
-import {MainPage} from "./headers";
+import {MainPage} from "./pages/MainPage";
 import './app.scss';
 
 const App = () => {
