+++ front-end/src/headers/HeaderContainer.js
@@ -2,7 +2,7 @@ import React from 'react';
 import HeaderContent from "./HeaderContent";
 
 const HeaderContainer = () => {
-    onclick = () => {
+    const onOpenBurger = () => {
         let menu__burger = document.querySelector('#menu__burger');
         let header__menu = document.querySelector('#header__menu');
         menu__burger.classList.toggle("active");
@@ -10,7 +10,7 @@ const HeaderContainer = () => {
     }
     return (
         <div className="header__container">
-            <div onClick={onclick} className="menu__burger burger" id="menu__burger">
+            <div onClick={onOpenBurger} className="menu__burger burger" id="menu__burger">
                 <span className="span__up"/>
                 <span className="span__middle"/>
                 <span className="span__down"/>
