+++ front-end/src/pages/MainPage/index.js
@@ -1,0 +1,5 @@
+import MainPage from "./MainPage";
+
+export  {
+    MainPage
+}
